Caregiver-Recommendation/Cloud-Functions/requirements.txt
functions-framework==3.*
pandas
matplotlib
numpy
scikit-learn
tensorflow
tensorflow_hub
google-cloud-storage